feat(host): delete unused blob providers
