+++ bin/host/src/lib.rs
@@ -2,13 +2,13 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
+pub mod blobs;
 pub mod cli;
 pub use cli::{init_tracing_subscriber, HostCli};
 
 pub mod fetcher;
 pub mod kv;
 pub mod preimage;
-pub mod providers;
 pub mod server;
 pub mod util;
