--- bin/host/src/providers/mod.rs
@@ -1,7 +1,0 @@
-//! Contains provider implementations for kona's host.
-
-pub mod blob;
-pub use blob::OnlineBlobProvider;
-
-pub mod beacon;
-pub use beacon::{BeaconClient, OnlineBeaconClient};
